+++ auth/src/main/kotlin/com/wwdt/auth/application/AuthApplication.kt
