+++ settings.gradle.kts
@@ -1,7 +1,3 @@
-pluginManagement {
-    repositories {
-        maven { url = uri("https://repo.spring.io/snapshot") }
-        gradlePluginPortal()
-    }
-}
 rootProject.name = "WWDT"
+include("shared_kernel")
+include("auth")
