+++ shared_kernel/src/main/kotlin/com/wwdt/shared_kernel/model/BaseEntity.kt
@@ -1,0 +1,26 @@
+package com.wwdt.shared_kernel.model
+
+import jakarta.persistence.Column
+import jakarta.persistence.EntityListeners
+import jakarta.persistence.MappedSuperclass
+import org.springframework.data.annotation.CreatedDate
+import org.springframework.data.annotation.LastModifiedDate
+import org.springframework.data.jpa.domain.support.AuditingEntityListener
+import java.time.LocalDateTime
+
+@MappedSuperclass
+@EntityListeners(AuditingEntityListener::class)
+class BaseEntity {
+
+    @CreatedDate
+    @Column(updatable = false)
+    // var 로 선언되어 있지만 CreatedDate 어노테이션 과 updatable = false 설정으로 인해 값이 변경되지 않음
+    var createdAt: LocalDateTime = LocalDateTime.now()
+
+    @LastModifiedDate
+    var updatedAt: LocalDateTime? = null
+
+    @Column
+    var deletedAt: LocalDateTime? = null
+
+}
