+++ shared_kernel/src/main/kotlin/com/wwdt/shared_kernel/utils/Extensions.kt
@@ -1,0 +1,2 @@
+package com.wwdt.shared_kernel.utils
+
