--- src/main/kotlin/com/wwdt/shared_kernal/utils/Extensions.kt
@@ -1,2 +1,0 @@
-package com.wwdt.shared_kernal.utils
-
