--- src/test/kotlin/com/wwdt/WwdtApplicationTests.kt
@@ -1,13 +1,0 @@
-package com.wwdt
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class WwdtApplicationTests {
-
-    @Test
-    fun contextLoads() {
-    }
-
-}
